Move classes to one package
