+++ collect_app/src/main/java/org/odk/collect/android/widgets/WidgetAnswer.kt
@@ -5,6 +5,7 @@ import androidx.compose.ui.Modifier
 import androidx.lifecycle.ViewModelProvider
 import org.javarosa.core.model.Constants
 import org.javarosa.form.api.FormEntryPrompt
+import org.odk.collect.android.widgets.arbitraryfile.ArbitraryFileWidgetAnswer
 import org.odk.collect.android.widgets.barcode.BarcodeWidgetAnswer
 import org.odk.collect.android.widgets.video.VideoWidgetAnswer
 
@@ -30,7 +31,13 @@ fun WidgetAnswer(
                 }
             }
             Constants.CONTROL_VIDEO_CAPTURE -> VideoWidgetAnswer(modifier, answer, viewModelProvider!!, onLongClick)
-            Constants.CONTROL_FILE_CAPTURE -> ArbitraryFileWidgetAnswer(modifier, answer, fontSize, viewModelProvider!!, onLongClick)
+            Constants.CONTROL_FILE_CAPTURE -> ArbitraryFileWidgetAnswer(
+                modifier,
+                answer,
+                fontSize,
+                viewModelProvider!!,
+                onLongClick
+            )
             else -> throw IllegalArgumentException("Unsupported control type: ${prompt.controlType}")
         }
     }
