+++ collect_app/src/main/java/org/odk/collect/android/widgets/arbitraryfile/ArbitraryFileWidgetContent.kt
@@ -1,29 +1,30 @@
-package org.odk.collect.android.widgets
+package org.odk.collect.android.widgets.arbitraryfile
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.OpenInNew
+import androidx.compose.material.icons.filled.AttachFile
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
+import org.odk.collect.android.widgets.WidgetIconButton
 import org.odk.collect.strings.R.string
 
 @Composable
-fun ExArbitraryFileWidgetContent(
+fun ArbitraryFileWidgetContent(
     readOnly: Boolean,
     fontSize: Int,
-    onLaunchClick: () -> Unit,
+    onChooseFileClick: () -> Unit,
     onLongClick: () -> Unit,
     widgetAnswer: @Composable () -> Unit
 ) {
     Column {
         if (!readOnly) {
             WidgetIconButton(
-                Icons.AutoMirrored.Filled.OpenInNew,
-                stringResource(string.launch_app),
+                Icons.Default.AttachFile,
+                stringResource(string.choose_file),
                 fontSize,
-                onLaunchClick,
+                onChooseFileClick,
                 onLongClick
             )
         }
@@ -34,9 +35,9 @@ fun ExArbitraryFileWidgetContent(
 
 @Preview
 @Composable
-private fun ExArbitraryFileWidgetContentPreview() {
+private fun ArbitraryFileWidgetContentPreview() {
     MaterialTheme {
-        ExArbitraryFileWidgetContent(
+        ArbitraryFileWidgetContent(
             false,
             10,
             {},
