+++ collect_app/src/main/java/org/odk/collect/android/widgets/arbitraryfile/ExArbitraryFileWidgetContent.kt
@@ -1,29 +1,30 @@
-package org.odk.collect.android.widgets
+package org.odk.collect.android.widgets.arbitraryfile
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.AttachFile
+import androidx.compose.material.icons.automirrored.filled.OpenInNew
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
+import org.odk.collect.android.widgets.WidgetIconButton
 import org.odk.collect.strings.R.string
 
 @Composable
-fun ArbitraryFileWidgetContent(
+fun ExArbitraryFileWidgetContent(
     readOnly: Boolean,
     fontSize: Int,
-    onChooseFileClick: () -> Unit,
+    onLaunchClick: () -> Unit,
     onLongClick: () -> Unit,
     widgetAnswer: @Composable () -> Unit
 ) {
     Column {
         if (!readOnly) {
             WidgetIconButton(
-                Icons.Default.AttachFile,
-                stringResource(string.choose_file),
+                Icons.AutoMirrored.Filled.OpenInNew,
+                stringResource(string.launch_app),
                 fontSize,
-                onChooseFileClick,
+                onLaunchClick,
                 onLongClick
             )
         }
@@ -34,9 +35,9 @@ fun ArbitraryFileWidgetContent(
 
 @Preview
 @Composable
-private fun ArbitraryFileWidgetContentPreview() {
+private fun ExArbitraryFileWidgetContentPreview() {
     MaterialTheme {
-        ArbitraryFileWidgetContent(
+        ExArbitraryFileWidgetContent(
             false,
             10,
             {},
